+++ opal/opal.rb
@@ -13,5 +13,6 @@ require 'corelib/dir'
 require 'corelib/file'
 require 'corelib/process'
 require 'corelib/random'
+require 'corelib/random/seedrandom.js'
 
 require 'corelib/unsupported'